Refactoring: Move additional to new factory
